added alert when timer ends, moved barbell info into a single component, moved note area into a single component
