+++ gym_app/src/type/BarbellInfo.type.tsx
@@ -1,0 +1,4 @@
+export type BarbellInfoType = {
+    haveBarbell: boolean;
+    totalWeight: number;
+}
